Add --overwrite option

+++ lib/mix/tasks/exfmt.all.ex
@@ -1,7 +1,9 @@
 defmodule Mix.Tasks.Exfmt.All do
   @moduledoc """
   Formats Elixir source code in specified directory recursively.
-      mix exfmt.all path/to/dir/
+  mix exfmt.all path/to/dir/
+  ## Command line options
+  * `--overwrite` Overwrite files that formatted by exfmt.
   """
 
   @shortdoc  "Format Elixir source code in specified directory."
@@ -21,15 +23,15 @@ defmodule Mix.Tasks.Exfmt.All do
   end
 
   def run(args) do
-    option_parser_options = [strict: [unsafe: :boolean]]
-    {_opts, [path], []} = OptionParser.parse(args, option_parser_options)
+    option_parser_options = [strict: [overwrite: :boolean]]
+    {opts, [path], []} = OptionParser.parse(args, option_parser_options)
     #TODO: Add code that checks path is a directory path.
     files = Path.join(path, "**/*.ex")
             |> Path.wildcard()
-    Enum.each(files, &format_dir(&1))
+    Enum.each(files, &format_dir(&1, opts))
   end
 
-  defp format_dir(path) do
+  defp format_dir(path, opts) do
     {:ok, source} = File.read(path)
     case Exfmt.format(source, 100) do
       {:ok, formatted} ->
@@ -39,8 +41,10 @@ defmodule Mix.Tasks.Exfmt.All do
           "#{path} does not formatted collectly!"
           |> red()
           |> IO.puts()
-          File.write!(path, formatted, [])
-          IO.write("#{path} is now formatted.")
+          if opts[:overwrite] do
+            File.write!(path, formatted, [])
+            IO.write("#{path} is now formatted.")
+          end
         end
       %SemanticsError{message: message} ->
         message
